// src/models/DSCFailedPaymentResponse.model.ts
export type DSCFailedPaymentListResponse = {
    _id: string
    propritorName: string
    srn: string
    email: string
    mobileNumber: string
    adhaarNumber: string
    address: string
    photoUrl: string
    adhaarCardPhotoUrl: string
    panCardPhotoUrl: string
    txnObjectId: string
    txnId: string
    uniqueTransactionId: string
    appliedFrom: string
    version: string
    acknowledgementPdf: string
    assignedTo: string
    assignedBy: string
    otherDocuments: any[]
    distributorCode: string
    distributorId: string
    totalPrice: number
    convinienceCharges: number
    basePrice: number
    baseCatAppied: string
    refundWalletAmountApplied: number
    rewardWalletAmountApplied: number
    acknowledgementNumber: string
    rejectionReason: string
    remark: string
    status: string
    txnStatus: string
    appliedByUserCurrentAddress: string
    appliedByType: string
    appliedById: string
    appliedOnDate: string
    appliedByName: string
    appliedByNumber: string
    appliedAsType: string
    assignedToName: string
    assignedToId: string
    assignedOnDate: string
    assignedByName: string
    assignedById: string
    verifiedByName: string
    verifiedById: string
    verifiedOnDate: string
    rejectedByName: string
    rejectedById: string
    rejectedOnDate: string
    generatedByName: string
    generatedById: string
    generatedOnDate: string
    completedByName: string
    completedById: string
    completedOnDate: string
    cancelledByName: string
    cancelledById: string
    cancelledOnDate: string
    moveToPendingByName: string
    moveToPendingById: string
    moveToPendingOnDate: string
    IsAgreedToTermsAndConditions: boolean
    retailerId: string
    retailerName: string
    retailerFirmName: string
    retailerFirmAddress: string
    retailerMobileNumber: string
    retailerPanNumber: string
    isDeleted: boolean
    createdAt: string
    updatedAt: string
    __v: number
  }